// src/lib/services/models/BillPaymentResponseChitStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { BillPaymentResponse } from './BillPaymentResponse';
import type { HttpStatusCode } from './HttpStatusCode';
export type BillPaymentResponseChitStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: BillPaymentResponse;
    statusCode?: HttpStatusCode;
    errors?: any;
};